Successfully fetched api data to table
